Float to double

+++ estore-api/src/test/java/com/estore/api/estoreapi/persistence/InventoryFileDAOTests.java
@@ -40,9 +40,9 @@ class InventoryFileDAOTests {
   void setupInventoryFileDAO() throws IOException {
       mockObjectMapper = mock(ObjectMapper.class);
       testBeefArray = new Beef[3];
-      testBeefArray[0] = new Beef(1, "Ribeye", (float)12.66, "A5", 299.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
-      testBeefArray[1] = new Beef(2,"Strip", (float)42.2, "C4", 13.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
-      testBeefArray[2] = new Beef(3, "Flank", (float)87, "C2", 9.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
+      testBeefArray[0] = new Beef(1, "Ribeye", 12.66, "A5", 299.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
+      testBeefArray[1] = new Beef(2,"Strip", 42.2, "C4", 13.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
+      testBeefArray[2] = new Beef(3, "Flank", 87, "C2", 9.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
 
       // When the object mapper is supposed to read from the file
       // the mock object mapper will return the beef array above
@@ -61,8 +61,9 @@ class InventoryFileDAOTests {
   }
 
   @Test
+
   void testUpdateBeef() throws IOException{
-    Beef updatedBeef = new Beef(1, "Ribeye", (float)1.34, "A5", 49.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
+    Beef updatedBeef = new Beef(1, "Ribeye", 1.34, "A5", 49.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
     Beef expectedBeef = new Beef(1, "Ribeye", 14, "A5", 49.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
 
     assertDoesNotThrow(() -> {
@@ -91,7 +92,7 @@ class InventoryFileDAOTests {
     Beef[] beefArray = inventoryFileDAO.findBeef("i");
 
     assertEquals(2, beefArray.length);
-    assertTrue(Arrays.equals(beefArray, new Beef[] {new Beef(1, "Ribeye", (float)12.66, "A5", 299.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ"), new Beef(2,"Strip", (float)42.2, "C4", 13.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ")}));
+    assertTrue(Arrays.equals(beefArray, new Beef[] {new Beef(1, "Ribeye", 12.66, "A5", 299.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ"), new Beef(2,"Strip", 42.2, "C4", 13.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ")}));
     assertTrue(Arrays.equals(inventoryFileDAO.findBeef("Chuck"), new Beef[] {}));
   }
 
@@ -105,13 +106,13 @@ class InventoryFileDAOTests {
 
   @Test
   void testCreateBeef(){
-    Beef beef = new Beef(4, "Skirt", (float)4.02, "C3", 8.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
+    Beef beef = new Beef(4, "Skirt", 4.02, "C3", 8.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
     assertDoesNotThrow(() -> {
       inventoryFileDAO.createBeef(beef);
     });
 
     assertEquals(beef, inventoryFileDAO.getBeef(4));
-    Beef duplicateBeef = new Beef(1, "Ribeye", (float)12.66, "A5", 299.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
+    Beef duplicateBeef = new Beef(1, "Ribeye", 12.66, "A5", 299.99, "https://www.youtube.com/watch?v=dQw4w9WgXcQ");
 
     assertDoesNotThrow(() -> {
       assertNull(inventoryFileDAO.createBeef(duplicateBeef));
